Improve add to-do to make it more aligned with regular todos look.
